Make the object cleaning code more general
The deque trick that was used to destroy old bullets only works if older
bullets are always destroyed before newer bullets. In the future, both
bullets and tanks will have health and they will need to be destroyed
when their health runs out, so a more general algorithm is needed. The
new algorithm scans all of the objects in each time step and replaces
the objects that need to be destroyed with objects from the end of the
vector.