Improve the management of objects
Objects now handle the destruction of their Box2D bodies. Object::step()
and Object::shouldDestroy() are now merged. The arena maintain the order
of the tanks since their barrels may overlap with other tanks.
Arena::bullets is renamed to Arena::objects since it will be used for
all non-tank objects in the future.